requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
flask==2.3.3
requests==2.31.0
plotly==5.18.0
pandas==2.1.1